Extract Code Actions
