+++ __init__.py
@@ -8,3 +8,4 @@ from .features.definition import *
 from .features.hover import *
 from .features.references import *
 from .features.signature_help import *
+from .features.code_actions import *
