+++ core/main.py
@@ -12,7 +12,7 @@ import sublime
 
 from .url import filename_to_uri, uri_to_filename
 from .protocol import (
-    Request, Notification, Point, Range, SymbolKind
+    Request, Notification, Range, SymbolKind
 )
 from .settings import (
     ClientConfig, settings, load_settings, unload_settings
@@ -29,9 +29,9 @@ from .clients import (
 )
 from .events import Events
 from .documents import (
-    initialize_document_sync, notify_did_open, get_position
+    initialize_document_sync, notify_did_open
 )
-from .diagnostics import handle_diagnostics, remove_diagnostics, get_line_diagnostics
+from .diagnostics import handle_diagnostics, remove_diagnostics
 
 
 symbol_kind_names = {
@@ -294,67 +294,6 @@ def start_server(server_binary_args, working_dir, env):
     except Exception as err:
         sublime.status_message("Failed to start LSP server {}".format(str(server_binary_args)))
         exception_log("Failed to start server", err)
-
-
-class LspCodeActionsCommand(sublime_plugin.TextCommand):
-    def is_enabled(self, event=None):
-        if is_supported_view(self.view):
-            client = client_for_view(self.view)
-            if client and client.has_capability('codeActionProvider'):
-                return True
-        return False
-
-    def run(self, edit, event=None):
-        client = client_for_view(self.view)
-        if client:
-            pos = get_position(self.view, event)
-            row, col = self.view.rowcol(pos)
-            line_diagnostics = get_line_diagnostics(self.view, pos)
-            params = {
-                "textDocument": {
-                    "uri": filename_to_uri(self.view.file_name())
-                },
-                "context": {
-                    "diagnostics": list(diagnostic.to_lsp() for diagnostic in line_diagnostics)
-                }
-            }
-            if len(line_diagnostics) > 0:
-                # TODO: merge ranges.
-                params["range"] = line_diagnostics[0].range.to_lsp()
-            else:
-                params["range"] = Range(Point(row, col), Point(row, col)).to_lsp()
-
-            if event:  # if right-clicked, set cursor to menu position
-                sel = self.view.sel()
-                sel.clear()
-                sel.add(sublime.Region(pos))
-
-            client.send_request(Request.codeAction(params), self.handle_codeaction_response)
-
-    def handle_codeaction_response(self, response):
-        titles = []
-        self.commands = response
-        for command in self.commands:
-            titles.append(
-                command.get('title'))  # TODO parse command and arguments
-        if len(self.commands) > 0:
-            self.view.show_popup_menu(titles, self.handle_select)
-        else:
-            self.view.show_popup('No actions available', sublime.HIDE_ON_MOUSE_MOVE_AWAY)
-
-    def handle_select(self, index):
-        if index > -1:
-            client = client_for_view(self.view)
-            if client:
-                client.send_request(
-                    Request.executeCommand(self.commands[index]),
-                    self.handle_command_response)
-
-    def handle_command_response(self, response):
-        pass
-
-    def want_event(self):
-        return True
 
 
 def apply_workspace_edit(window, params):
