+++ features/code_actions.py
@@ -1,0 +1,77 @@
+import sublime
+import sublime_plugin
+
+try:
+    from typing import Any, List
+    assert Any and List
+except ImportError:
+    pass
+
+
+from ..core.clients import client_for_view
+from ..core.configurations import is_supported_view
+from ..core.protocol import Request, Range, Point
+from ..core.documents import get_position
+from ..core.diagnostics import get_line_diagnostics
+from ..core.url import filename_to_uri
+
+
+class LspCodeActionsCommand(sublime_plugin.TextCommand):
+    def is_enabled(self, event=None):
+        if is_supported_view(self.view):
+            client = client_for_view(self.view)
+            if client and client.has_capability('codeActionProvider'):
+                return True
+        return False
+
+    def run(self, edit, event=None):
+        client = client_for_view(self.view)
+        if client:
+            pos = get_position(self.view, event)
+            row, col = self.view.rowcol(pos)
+            line_diagnostics = get_line_diagnostics(self.view, pos)
+            params = {
+                "textDocument": {
+                    "uri": filename_to_uri(self.view.file_name())
+                },
+                "context": {
+                    "diagnostics": list(diagnostic.to_lsp() for diagnostic in line_diagnostics)
+                }
+            }
+            if len(line_diagnostics) > 0:
+                # TODO: merge ranges.
+                params["range"] = line_diagnostics[0].range.to_lsp()
+            else:
+                params["range"] = Range(Point(row, col), Point(row, col)).to_lsp()
+
+            if event:  # if right-clicked, set cursor to menu position
+                sel = self.view.sel()
+                sel.clear()
+                sel.add(sublime.Region(pos))
+
+            client.send_request(Request.codeAction(params), self.handle_codeaction_response)
+
+    def handle_codeaction_response(self, response):
+        titles = []
+        self.commands = response
+        for command in self.commands:
+            titles.append(
+                command.get('title'))  # TODO parse command and arguments
+        if len(self.commands) > 0:
+            self.view.show_popup_menu(titles, self.handle_select)
+        else:
+            self.view.show_popup('No actions available', sublime.HIDE_ON_MOUSE_MOVE_AWAY)
+
+    def handle_select(self, index):
+        if index > -1:
+            client = client_for_view(self.view)
+            if client:
+                client.send_request(
+                    Request.executeCommand(self.commands[index]),
+                    self.handle_command_response)
+
+    def handle_command_response(self, response):
+        pass
+
+    def want_event(self):
+        return True
